+++ scripts/generate-website-yaml.py
@@ -36,26 +36,6 @@ if __name__ == "__main__":
         with engine.connect() as connection:
             # Generate deployment table
             df_depl = config.get_deployment_table(connection)
-            x = df_depl.copy(deep=True)
-            x = x.drop(["Dates", "Sensors"], axis=1)
-
-            # Write Deployments table to fleet status
-            wk_name = "Deployments-Database"
-            logging.info("Updating the Fleet Status %s sheet", wk_name)
-            x = x.fillna("").rename({"Glider_Deployment_ID": "Deployment_ID"})
-            gc = gspread.oauth()  # type: ignore
-            sh = gc.open("Fleet Status")
-            wk = sh.worksheet(wk_name)
-            wk.update([x.columns.values.tolist()] + x.values.tolist())
-
-            # # Update data validation formatting automatically..
-            # from gspread.utils import ValidationConditionType
-            # wk.add_validation(
-            #     f'F2:L{1+x.shape[0]}',
-            #     ValidationConditionType.one_of_list,
-            #     ['TRUE', 'FALSE'],
-            #     showCustomUi=True
-            # )
 
             # Make website yaml
             config.make_website_yaml(df_depl, yaml_path)